+++ Assets/Items/Wheeler/WheelerBody.mat
@@ -114,7 +114,7 @@ Material:
     - _GlossMapScale: 1
     - _Glossiness: 0.5
     - _GlossyReflections: 1
-    - _Metallic: 0.378
+    - _Metallic: 0.549
     - _Mode: 0
     - _OcclusionStrength: 1
     - _Parallax: 0.02
