--- Assets/Race Courses/Course 5/FramedWindow.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 67c9f65661d4bf64380f908c606d6b7b
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
