--- Assets/Race Courses/Course 5/PoppableGlass.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1ac8110df105fbe4bafea165858eea2c
