--- Assets/Race Courses/Course 5/TriangleWindow.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 24561cccf2934ac48a4ef0b29addf79e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
