+++ Assets/Race Courses/Course 5/Windows/Glass.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ce3d04b5231ee9b4c9f73ed73f46e2d4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
