+++ Assets/Race Courses/Course 5/Windows/Glass/BreakableGlass.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 63b340be89a01ab459acb5e61a0142ab
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
