+++ Assets/Race Courses/Course 5/Windows/Glass/Glass.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ca17f87973c05ee46aa180f2419014e5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
