+++ Assets/Race Courses/Course 5/Windows/TriangleWindow.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 24561cccf2934ac48a4ef0b29addf79e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
