+++ Assets/Race Courses/Course 5/Windows/WindowFrame.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 765a4efa151eac4458f5d107e514f97a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
