+++ Assets/Race Courses/Course 5/mall_materials/Threshold.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cd17f39ddebd185488ff8837907d355c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
