--- Assets/Race Courses/Glass.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ce3d04b5231ee9b4c9f73ed73f46e2d4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
