--- Assets/Race Courses/Glass/BreakableGlass.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 63b340be89a01ab459acb5e61a0142ab
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
